+++ swing/common-ui/src/main/java/module-info.java
@@ -71,6 +71,7 @@ module is.codion.swing.common.ui {
 	exports is.codion.swing.common.ui.key;
 	exports is.codion.swing.common.ui.laf;
 	exports is.codion.swing.common.ui.layout;
+	exports is.codion.swing.common.ui.transfer;
 	exports is.codion.swing.common.ui.window;
 
 	uses is.codion.swing.common.ui.laf.LookAndFeelProvider;