+++ changelog.md
@@ -38,6 +38,7 @@ Codion Change Log
 - Sizes moved to component package.
 - Colors moved to color package.
 - Cursors moved to cursor package.
+- FileTransferHandler moved to transfer package.
 
 ## 0.18.39
 ### is.codion.common.core